% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/cross_validation.R
\name{kfold_validation}
\alias{kfold_validation}
\title{perform kfold validation for a model object}
\usage{
kfold_validation(object, ...)
}
\arguments{
\item{object}{a model object}

\item{k}{define the k-fold validation}
}
\description{
perform kfold validation for a model object
}
\details{
perform kfold validation for a model object
}
\examples{
utils::data(anorexia, package = "MASS")
anorex.1 <- glm(Postwt ~ Prewt + Treat + offset(Prewt), family = gaussian, data = anorexia)
kfold_validation(anorex.1, 10)
}

